Update test format to modern qunit
